requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
requests
pycoingecko
coinpaprika
pytest
services
tenacity
prometheus-fastapi-instrumentator